Adding a sample eCommerce
NOTE: This code was part of a Interview process but cannot indicate the source.

I just polished and will do to my latest coding standards.

By far the best part is the LIST validation of the request payload in the ProductRequest and validated when touching the ProductController.addProduct() method.
Also, the way to validate a PathVaiable (but need more testing).